worked around inheritance problem
the inheritance problem persists, because when we inherit from an object
we copy over the prototype of the base object, and because the prototype
of object contained `__init__` we found that before we found `__init__`
in the dict of the base object. mro looks in the prototype of an object
when doing a typelookup.

To mitigate it I moved the `__init__` of object to the `$d` dict. It
solves the problem, but it doesn't solve the larger problem. We will
have to revisit this, to maybe not have the mro look in the prototype or
maybe to not copy python functions from the parent object.